+++ src/components/DateInput/index.jsx
@@ -4,14 +4,12 @@ import moment from 'moment';
 
 import DayPicker from 'react-day-picker';
 
-import ActionsArrowLeftIcon from 'lib/icons/react/ActionsArrowLeftIcon';
-import ActionsArrowRightIcon from 'lib/icons/react/ActionsArrowRightIcon';
 import DatePickerIcon from 'lib/icons/react/DatePickerIcon';
 import Popover from 'components/Popover';
-import Flex from 'components/Flex';
-import FlexItem from 'components/FlexItem';
-import IconButton from 'components/IconButton';
 import TextInput from 'components/TextInput';
+import YearAndMonthSelector from './YearAndMonthSelector';
+import NavArrows from './NavArrows';
+import { getYearRange, isValidUserDate } from './util';
 
 // Is this the correct way to localize dates? No, it is not.
 // Fortunately, this is display-only.
@@ -23,144 +21,6 @@ export const DATE_FORMAT_MAP = {
   DMY: 'DD/MM/YYYY',
   YMD: 'YYYY/MM/DD', // ISO-8601, the best but least used standard
 };
-
-// capture groups for M and D work with or without a leading zero.
-// Y component requires 4 digits for all formats.
-const DATE_PATTERN_MAP = {
-  MDY: /^(0?[1-9]|1[0-2])[/](0?[1-9]|[12]\d|3[01])[/]\d{4}$/,
-  DMY: /^(0?[1-9]|[12]\d|3[01])[/](0?[1-9]|1[0-2])[/]\d{4}$/,
-  YMD: /^\d{4}[/](0?[1-9]|1[0-2])[/](0?[1-9]|[12]\d|3[01])$/,
-};
-
-/**
- * Returns a range of years to show in date picker. The range will expand to include
- * the selected date, if a date has been selected.
- *
- * @param {Number} currentYear YYYY
- * @param {Number} pastYears number of prior years in range
- * @param {Number} futureYears number of future years in range
- * @param {Date} minDate lower bound of selectable dates
- * @param {Date} maxDate upper bound of selectable dates
- * @param {Date} selectedDate currently selected date from DateInput
- * @returns {Object} { startYear: YYYY, endYear: YYYY }
- */
-export const getYearRange = (
-  currentYear,
-  pastYears,
-  futureYears,
-  minDate,
-  maxDate,
-  selectedDate
-) => {
-  // use min/max dates if specified
-  // fall back relative past/future years from current date
-  let startYear =
-    (minDate && minDate.getFullYear()) ||
-    new Date(currentYear - pastYears, 0).getFullYear();
-  let endYear =
-    (maxDate && maxDate.getFullYear() + 1) ||
-    new Date(currentYear + futureYears + 1, 11).getFullYear();
-  const selectedYear = selectedDate instanceof Date && selectedDate.getFullYear();
-
-  // Expand range to include selected year if out of range
-  if (selectedYear && selectedYear < startYear) {
-    startYear = selectedYear;
-  }
-  if (selectedYear && selectedYear > endYear) {
-    endYear = selectedYear + 1;
-  }
-
-  return { startYear, endYear };
-};
-
-/**
- * Checks user-entered date strings for validity and completeness.
- *
- * moment eagerly parses dates (starting with a single number!), so
- * we must wait for a user to finish typing something that looks like
- * a full date.
- *
- * We use patterns that accept a date string with or without a leading zero.
- *
- * @param {String} inputValue
- * @param {String} dateFormat (MDY/DMY/YMD)
- * @returns {Boolean}
- */
-export const isValidUserDate = (inputValue, dateFormat) =>
-  DATE_PATTERN_MAP[dateFormat || 'MDY'].test(inputValue);
-
-/* eslint-disable react/prop-types */
-/**
- * Private component for DateInput that renders Month/Year select inputs.
- * @param {Object} props react props
- * @returns {ReactElement}
- */
-const YearAndMonthSelector = ({ date, localeUtils, onChange, startYear, endYear }) => {
-  const months = localeUtils.getMonths();
-
-  const years = Array.from({ length: endYear - startYear }, (year, i) => i + startYear);
-
-  const handleChange = ({
-    target: {
-      form: { year, month },
-    },
-  }) => {
-    onChange(new Date(year.value, month.value));
-  };
-
-  return (
-    <form className="DayPicker-Caption border--bottom">
-      <div className="padding--bottom--half">
-        <select
-          name="month"
-          onChange={handleChange}
-          value={date.getMonth()}
-          className="margin--right--half"
-        >
-          {months.map((month, i) => (
-            <option key={month} value={i}>
-              {month}
-            </option>
-          ))}
-        </select>
-        <select name="year" onChange={handleChange} value={date.getFullYear()}>
-          {years.map((year) => (
-            <option key={year} value={year}>
-              {year}
-            </option>
-          ))}
-        </select>
-      </div>
-    </form>
-  );
-};
-
-/**
- * Private component for DateInput that renders prev/next arrows.
- * @param {Object} props react props
- * @returns {ReactElement}
- */
-const NavArrows = ({ onPreviousClick, onNextClick }) => (
-  <div className="fdsDateInput-navArrows alignChild--right--center">
-    <Flex noGutters>
-      <FlexItem shrink>
-        <IconButton
-          onClick={() => onPreviousClick()}
-          Icon={ActionsArrowLeftIcon}
-          label="Previous Month"
-        />
-      </FlexItem>
-      <FlexItem shrink>
-        <IconButton
-          onClick={() => onNextClick()}
-          Icon={ActionsArrowRightIcon}
-          label="Next Month"
-        />
-      </FlexItem>
-    </Flex>
-  </div>
-);
-/* eslint-enable react/prop-types */
 
 /**
  * @param {Object} props react props
